Cast model to bfloat16.

+++ tests/test_modeling_common.py
@@ -3484,7 +3484,7 @@ class ModelTesterMixin:
             model = model_class(config)
 
             model.to(torch_device)
-            model.half()
+            model.to(torch.bfloat16)
             dummy_input = inputs_dict[model.main_input_name][:1]
             if dummy_input.dtype in [torch.float32, torch.float16]:
                 dummy_input = dummy_input.to(torch.bfloat16)
